Make computeSets reactive instead of state
